+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changes for croud
 Unreleased
 ==========
 
+- Added new ``croud me edit`` command to change your own email address.
+
 0.19.1 - 2019/11/18
 ===================